Commit all new files
